+++ addon/objects/uml-primitives/fd-uml-primitive.js
@@ -204,11 +204,6 @@ joint.highlighters.strokeAndButtons = {
     }
 
     if (cellView.getSizeChangers instanceof Function) {
-      if (cellView.el.dataset.type !== 'flexberry.uml.Partition') {
-        const parentGroup = document.querySelector('g[joint-selector="cells"]');
-        const selectedСlass = parentGroup.querySelector(`#${cellView.id}`);
-        parentGroup.appendChild(selectedСlass);
-      }
       this.addSizeChangers(cellView, id);
     }